# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /arch/analyze:
    post:
      operationId: analyzeArchDescription
      tags: [arch]
      summary: Analyze a free-text agent description and infer architecture parameters
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArchAnalyzeBody"
      responses:
        "200":
          description: Inferred architecture parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArchAnalyzeResult"
  /arch/generate:
    post:
      operationId: generateArchSpec
      tags: [arch]
      summary: Generate a full architecture spec as an SSE stream
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArchGenerateBody"
      responses:
        "200":
          description: SSE stream of architecture spec chunks
          content:
            text/event-stream: {}
components:
  schemas:
...
    ArchAnalyzeBody:
      type: object
      properties:
        description:
          type: string
      required:
        - description
    ArchAnalyzeResult:
      type: object
      properties:
        topology:
          type: string
        autonomy:
          type: string
        memory:
          type: string
        failure:
          type: string
        platform:
          type: string
        scaleRuns:
          type: string
        scaleConcurrent:
          type: string
        greenfield:
          type: string
        modelPref:
          type: string
        budget:
          type: string
        security:
          type: array
          items:
            type: string
        inferredTools:
          type: array
          items:
            type: string
        gaps:
          type: array
          items:
            type: string
        summary:
          type: string
      required:
        - topology
        - autonomy
        - memory
        - failure
        - platform
        - scaleRuns
        - scaleConcurrent
        - greenfield
        - modelPref
        - budget
        - security
        - inferredTools
        - gaps
        - summary
    ArchGenerateBody:
      type: object
      properties:
        description:
          type: string
        answers:
          type: object
          additionalProperties:
            type: string
        security:
          type: array
          items:
            type: string
        tools:
          type: array
          items:
            type: string
        constraints:
          type: string
        customAgents:
          type: string
        legacyDesc:
          type: string
        summary:
          type: string
      required:
        - description
        - answers
        - security
        - tools