# lib/api-zod/src/generated/types/archAnalyzeResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ArchAnalyzeResult {
  topology: string;
  autonomy: string;
  memory: string;
  failure: string;
  platform: string;
  scaleRuns: string;
  scaleConcurrent: string;
  greenfield: string;
  modelPref: string;
  budget: string;
  security: string[];
  inferredTools: string[];
  gaps: string[];
  summary: string;
}