# lib/api-zod/src/generated/types/archGenerateBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ArchGenerateBodyAnswers } from "./archGenerateBodyAnswers";

export interface ArchGenerateBody {
  description: string;
  answers: ArchGenerateBodyAnswers;
  security: string[];
  tools: string[];
  constraints?: string;
  customAgents?: string;
  legacyDesc?: string;
  summary?: string;
}